requirements_cuda.txt
altgraph==0.17.4
attrs==23.1.0
auto-py-to-exe==2.42.0
backcall==0.2.0
bottle==0.12.25
bottle-websocket==0.2.9
cdist==7.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
circle-fit==0.1.3
click==8.1.3
colorama==0.4.6
ConfigArgParse==1.5.3
cycler==0.11.0
Cython==0.29.34
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
Eel==0.16.0
entrypoints==0.4
fastjsonschema==2.16.3
Flask==2.2.5
fonttools==4.38.0
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==3.0.2
hdbscan==0.8.37
idna==3.4
imageio==2.28.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
laspy==2.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
nbformat==5.7.0
nest-asyncio==1.5.6
numpy==1.21.6
open3d==0.17.0
packaging==23.1
pandas==1.3.5
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
plotly==5.14.1
pooch==1.7.0
prompt-toolkit==3.0.38
psutil==5.9.5
pycobra==0.2.5
pycparser==2.21
Pygments==2.15.1
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.10
pyntcloud==0.3.1
pyparsing==3.0.9
pypcd @ git+https://github.com/DanielPollithy/pypcd.git@88ab8c98ab81dd620bf4b22d965b37457aab78f8
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyrsistent==0.19.3
pyshp==2.3.1
python-dateutil==2.8.2
python-lzf==0.2.4
pytz==2023.3
pyvista==0.38.5
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.0.2
requests==2.28.2
scikit-learn==1.0.2
scipy==1.7.3
scooby==0.7.1
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
tenacity==8.2.2
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
vtk==9.2.6
wcwidth==0.2.6
Werkzeug==2.2.3
whichcraft==0.6.1
widgetsnbextension==4.0.7
zipp==3.15.0
zope.event==5.0
zope.interface==6.1
numba==0.55.2